refactor: split public types into separate files, remove SEP-990 refs from XML docs
